Feature - Spatie permissions (#8)
* Add spatie permissions package

* Add the permissions syncing func (thanks to @dwaard)

* Add policy and implement it throughout the service layer and the livewire components

* Update the unit tests

* Fix phpstan and lint issues

* Fix coding style using Pint

+++ app/Livewire/Pages/Internal/Partner/Delete.php
@@ -32,13 +32,18 @@ class Delete extends Component
     public function setUpModal(int $partnerId): void
     {
         $this->reset();
+
         $this->partnerId = $partnerId;
         $this->partner = $this->partnerService->getPartner($partnerId);
+        $this->authorize('delete', $this->partner);
+
         $this->deleteIsOpen = true;
     }
 
     public function confirm(): void
     {
+        $this->authorize('delete', $this->partner);
+
         try {
             $this->partnerService->deletePartner($this->partnerId);
             $this->cleanUpSuccessfully();
